big refactoring. - move implementation of dev-only modules under a __DEV__ block. this ensures good DCE result for webpack/browserify users and FB's packager. - an explicit 'Dev' postfix is added to their module names
modules renamed:
- ReactElementValidator.js -> ReactElementValidatorDev.js
- ReactComponentTreeDevtool.js -> ReactComponentTreeDevtoolDev.js
- validateDOMNesting.js -> validateDOMNestingDev.js
- ReactDOMDebugTool.js -> ReactDOMDebugToolDev.js
- ReactDOMNullInputValuePropDevtool.js -> ReactDOMNullInputValuePropDevtoolDev.js
- ReactDOMUnknownPropertyDevtool.js -> ReactDOMUnknownPropertyDevtoolDev.js
- ReactDebugTool.js -> ReactDebugToolDev.js
- ReactInstrumentation.js -> ReactInstrumentationDev.js
- ReactPerf.js -> ReactPerfDev.js
- ReactChildrenMutationWarningDevtool.js -> ReactChildrenMutationWarningDevtoolDev.js
- ReactHostOperationHistoryDevtool.js -> ReactHostOperationHistoryDevtoolDev.js
- ReactInvalidSetStateWarningDevTool.js -> ReactInvalidSetStateWarningDevToolDev.js
- canDefineProperty.js -> canDefinePropertyDev.js
