+++ src/isomorphic/classic/element/ReactDOMFactories.js
@@ -12,7 +12,7 @@
 'use strict';
 
 var ReactElement = require('ReactElement');
-var ReactElementValidator = require('ReactElementValidator');
+var ReactElementValidatorDev = require('ReactElementValidatorDev');
 
 var mapObject = require('mapObject');
 
@@ -24,7 +24,7 @@ var mapObject = require('mapObject');
  */
 function createDOMFactory(tag) {
   if (__DEV__) {
-    return ReactElementValidator.createFactory(tag);
+    return ReactElementValidatorDev.createFactory(tag);
   }
   return ReactElement.createFactory(tag);
 }
