+++ src/renderers/dom/client/ReactReconcileTransaction.js
@@ -15,7 +15,7 @@ var CallbackQueue = require('CallbackQueue');
 var PooledClass = require('PooledClass');
 var ReactBrowserEventEmitter = require('ReactBrowserEventEmitter');
 var ReactInputSelection = require('ReactInputSelection');
-var ReactInstrumentation = require('ReactInstrumentation');
+var ReactInstrumentationDev = require('ReactInstrumentationDev');
 var Transaction = require('Transaction');
 var ReactUpdateQueue = require('ReactUpdateQueue');
 
@@ -94,8 +94,8 @@ var TRANSACTION_WRAPPERS = [
 
 if (__DEV__) {
   TRANSACTION_WRAPPERS.push({
-    initialize: ReactInstrumentation.debugTool.onBeginFlush,
-    close: ReactInstrumentation.debugTool.onEndFlush,
+    initialize: ReactInstrumentationDev.debugTool.onBeginFlush,
+    close: ReactInstrumentationDev.debugTool.onEndFlush,
   });
 }
