+++ src/renderers/dom/shared/__tests__/ReactDOMTextComponentDev-test.js
