+++ src/renderers/native/ReactNativeReconcileTransaction.js
@@ -14,7 +14,7 @@
 var CallbackQueue = require('CallbackQueue');
 var PooledClass = require('PooledClass');
 var Transaction = require('Transaction');
-var ReactInstrumentation = require('ReactInstrumentation');
+var ReactInstrumentationDev = require('ReactInstrumentationDev');
 var ReactUpdateQueue = require('ReactUpdateQueue');
 
 /**
@@ -46,8 +46,8 @@ var TRANSACTION_WRAPPERS = [ON_DOM_READY_QUEUEING];
 
 if (__DEV__) {
   TRANSACTION_WRAPPERS.push({
-    initialize: ReactInstrumentation.debugTool.onBeginFlush,
-    close: ReactInstrumentation.debugTool.onEndFlush,
+    initialize: ReactInstrumentationDev.debugTool.onBeginFlush,
+    close: ReactInstrumentationDev.debugTool.onEndFlush,
   });
 }
